refactor(github_search): separate state and view
