+++ lib/state/github_search_page/search_bar_visible_state.dart
@@ -1,0 +1,3 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final searchBarVisibleProvider = StateProvider<bool>((ref) => true);
